% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_design_matrix}
\alias{create_design_matrix}
\title{Create design matrix for time series analysis}
\usage{
create_design_matrix(X = NULL, y, lags_x = NULL, lags_y = lags_x)
}
\arguments{
\item{X}{a matrix or data frame of covariates}

\item{y}{a vector of response variable}

\item{lags_x}{number of lags for the covariates, defaults to NULL}

\item{lags_y}{number of lags for the response variable, defaults to the same value as lags_x}
}
\value{
X_new a matrix of the design matrix
}
\description{
This function creates a design matrix from a given set of covariates (X) and a response variable (y). The design matrix includes lagged values of X and y.
}
