% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.fit_flexcode_timeseries}
\alias{plot.fit_flexcode_timeseries}
\title{plot_fit_flexcode_timeseries}
\usage{
\method{plot}{fit_flexcode_timeseries}(fit, X_new, y_new, predictionBandProb = TRUE, y_train = NULL)
}
\arguments{
\item{fit}{An object of class fit_flexcode_timeseries.}

\item{X_new}{A matrix containing the covariates for each new instance since the last observation used for training.}

\item{y_new}{A vector containing the response for each new instance since the last observation used for training. It has one last element than the number of rows of X_new. y_new must be NA if nrow(X_new)=1.}

\item{predictionBandProb}{A logical indicating whether to return the prediction band probability.}

\item{y_train}{A vector containing the response for the training observations. If y_train is not provided, only the prediction region for the new observations will be plotted.}
}
\value{
A plot of the estimated density for the new observations.
}
\description{
plot_fit_flexcode_timeseries takes an object fit, a matrix X_new, a vector y_new, a logical predictionBandProb, and a vector y_train as input and returns a plot of the estimated density for the new observations.
}
