% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{predict.fit_flexcode_timeseries}
\alias{predict.fit_flexcode_timeseries}
\title{predict_fit_flexcode_timeseries}
\usage{
\method{predict}{fit_flexcode_timeseries}(fit, X_new = NULL, y_new = NA, predictionBandProb = FALSE)
}
\arguments{
\item{fit}{An object of class fit_flexcode_timeseries.}

\item{X_new}{A matrix containing the covariates for each new instance since the last observation used for training.}

\item{y_new}{A vector containing the response for each new instance since the last observation used for training. It has one last element than the number of rows of X_new. y_new must be NA if nrow(X_new)=1.}

\item{predictionBandProb}{A logical indicating whether to return the prediction band probability.}
}
\value{
A list with the predicted values and the estimated density.
}
\description{
predict_fit_flexcode_timeseries takes an object fit, a matrix X_new, a vector y_new, and a logical predictionBandProb as input. It returns the fitted model and the estimated density for the new observations.
}
\examples{
fit <- fit_flexcode_timeseries(X_train, y_train)
predict_fit_flexcode_timeseries(fit, X_new, y_new, predictionBandProb = TRUE)
}
